Refactory/Fix: Charts
Some refactory/clean up, but also:
In the old OpenAstro code there was a -4 hardcoded to remove the last 4
points/planets in a chart, probably the original writer had in mind that
the last 4 elements should always be the Axes.
I created a logic to check if some points are axes then prevent drawing
it in the transit circle.